subscription-analytics/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
google-generativeai==0.3.2
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0
aiofiles==23.2.1
pydantic==2.5.0
httpx==0.25.2
openpyxl==3.1.2
